move history logic to history component
